+++ src/main/java/com/proyecto/web/product/dto/response/ProductResponse.java
@@ -1,13 +1,15 @@
 package com.proyecto.web.product.dto.response;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
 
         Long id,
         String code,
         String name,
         String category,
-        Double purchasePrice,
-        Double salePrice,
+        BigDecimal purchasePrice,
+        BigDecimal salePrice,
         Integer stock,
         SupplierShortResponse supplier
 ) {
